Prod release version 5

+++ server/artwork.js
@@ -81,21 +81,23 @@ export const getArtwork = (req, res) => {
     .catch((e) => console.error(e));
 };
 
-// function for DELETE endpoint
+// functions for DELETE endpoint
 
 export const deleteArtwork = (req, res) => {
   const id = req.params.id;
   db.query("SELECT * FROM artwork WHERE id=$1", [id])
     .then((result) => {
       if (result.rows.length > 0) {
-        const fileLink = result.rows[0].content_link
-		const fileKey = fileLink.slice(fileLink.indexOf("amazonaws.com/") + 14);
-        console.log(fileKey);
+        const fileKey = result.rows[0].content_link;
         db.query("DELETE FROM artwork WHERE id=$1", [id])
           .then(() => {
-            deleteFile(fileKey).then(() =>
-              res.json({ success: `Item ${id} deleted!` })
-            );
+            if (fileKey.length > 0) {
+              deleteFile(fileKey).then(() =>
+                res.json({ success: `Item ${id} deleted!` })
+              );
+            } else {
+              res.json({ success: `Item ${id} deleted!` });
+            }
           })
           .catch((e) => console.error(e));
       } else {
